requirements.txt
# 基础框架
fastapi==0.95.0
uvicorn[standard]==0.21.1
websockets==11.0.3

# 数据库与缓存
asyncpg==0.27.0
aioredis==2.0.1
sqlalchemy[asyncio]==2.0.10
alembic==1.10.2

# 深度学习核心
torch==1.13.1+cu117 --extra-index-url https://download.pytorch.org/whl/cu117
transformers==4.28.1
bitsandbytes==0.38.1
deepseek-sdk==2.4.2

# 数据处理
numpy==1.24.2
pandas==1.5.3
orjson==3.8.7

# 安全与认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# 系统工具
loguru==0.6.0
python-dotenv==0.21.1
uvloop==0.17.0

# 监控与性能
prometheus-client==0.16.0
sentry-sdk==1.15.0

# 生产服务器
gunicorn==20.1.0
